Replace TCI (IRT) with simple average score
The IRT-based Telco Capability Index was over-engineered for our needs.
Replace it with a simple arithmetic mean of all 5 benchmark scores.

- Delete .github/scripts/tci/ module (IRT fitter, calculator, types)
- Delete calculate_tci.py, create calculate_avg_score.py
- Remove numpy/scipy deps from workflow (no longer needed)
- Drop legacy 'tci' column from HF dataset, add 'avg_score'
