--- .github/scripts/tci/__init__.py
@@ -1,25 +1,0 @@
-"""TCI calculation module for GitHub Actions."""
-
-from .tci_calculator import (
-    LeaderboardEntry,
-    calculate_all_tci,
-    calculate_error,
-    calculate_tci,
-    sort_by_tci,
-)
-from .irt_fitter import (
-    BENCHMARKS,
-    IRTParameters,
-    fit_irt_parameters,
-)
-
-__all__ = [
-    "LeaderboardEntry",
-    "calculate_all_tci",
-    "calculate_error",
-    "calculate_tci",
-    "sort_by_tci",
-    "BENCHMARKS",
-    "IRTParameters",
-    "fit_irt_parameters",
-]
